+++ CHANGELOG.md
@@ -63,3 +63,4 @@ from fuseki. Remove temp file processing for loading og graph -- load directly. 
 * **2015-04-22** Begin work on person ingest.  Too hard.  Without position_data.csv
 * **2015-04-29** Work on make_person_update_data in uf_examples.
 * **2015-05-01** Add salary plan enum to uf_examples
+* **2015-05-02** Progress on make_person_update_data